chore: extracted all DateRangePicker logic from event-types/[type].tsx
